requirements.txt
APScheduler==3.6.3
cachetools==4.2.2
certifi==2021.10.8
dnspython==2.2.1
lxml==4.8.0
pymongo==4.0.2
PySocks==1.7.1
python-docx==0.8.11
python-telegram-bot==13.11
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
six==1.16.0
tornado==6.1
tzdata==2021.5
tzlocal==4.1
